Move components away from Layout folder
