+++ packages/client/src/components/StyledErrorBoundary.tsx
@@ -14,7 +14,7 @@ import * as React from 'react'
 import * as Sentry from '@sentry/browser'
 import styled from '@client/styledComponents'
 import { injectIntl, WrappedComponentProps as IntlShapeProps } from 'react-intl'
-import { PageWrapper } from '@opencrvs/components/lib/layout/PageWrapper'
+import { PageWrapper } from '@opencrvs/components/lib/PageWrapper'
 import { TertiaryButton } from '@opencrvs/components/lib/buttons'
 import { Box } from '@opencrvs/components/lib/interface/Box'
 import { errorMessages, buttonMessages } from '@client/i18n/messages'
