+++ packages/client/src/views/EventInfo/EventInfo.tsx
@@ -11,7 +11,7 @@
  */
 import * as React from 'react'
 import { RouteComponentProps } from 'react-router'
-import { Container, BodyContent } from '@opencrvs/components/lib/layout'
+import { Container, BodyContent } from '@opencrvs/components/lib/Content'
 import {
   WrappedComponentProps as IntlShapeProps,
   MessageDescriptor,
