+++ packages/client/src/views/OfficeHome/Outbox.tsx
@@ -10,7 +10,7 @@
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
 import * as React from 'react'
-import { BodyContent } from '@opencrvs/components/lib/layout'
+import { BodyContent } from '@opencrvs/components/lib/Content'
 import { Spinner } from '@opencrvs/components/lib/interface'
 import {
   Workqueue,
