+++ packages/client/src/views/SysAdmin/SysAdminContentWrapper.tsx
@@ -13,7 +13,7 @@
 import { MarginedHeader } from '@client/components/interface/Header/Header'
 import { CircleButton } from '@opencrvs/components/lib/buttons'
 import { BackArrowDeepBlue } from '@opencrvs/components/lib/icons'
-import { BodyContent, Container } from '@opencrvs/components/lib/layout'
+import { BodyContent, Container } from '@opencrvs/components/lib/Content'
 import { LoadingGrey } from '@opencrvs/components/lib/interface'
 import * as React from 'react'
 import styled from 'styled-components'
