+++ packages/components/src/components/PageWrapper/index.ts
@@ -9,5 +9,4 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './Content'
 export * from './PageWrapper'
